examples/native_computer_use_demo/requirements.txt
openai>=2.26.0,<3
Pillow>=11.3.0,<12
PyAutoGUI>=0.9.54,<1
